chore: delete @guildedjs/embeds and move to guilded.js
