--- packages/embeds/lib/index.ts
@@ -1,1 +1,0 @@
-export * from "./Embed";
