--- packages/embeds/package.json
@@ -1,66 +1,0 @@
-{
-    "name": "@guildedjs/embeds",
-    "version": "0.2.0",
-    "description": "Library-agnostic guilded embed builder utility for the guilded.gg API",
-    "author": "Zaid \"Nico\" <contact@nico.engineer>",
-    "license": "MIT",
-    "main": "dist/index.js",
-    "types": "types/index.d.ts",
-    "typedocMain": "lib/index.ts",
-    "scripts": {
-        "build": "npx tsc && npx gen-esm-wrapper . ./dist/index.mjs",
-        "prepublishOnly": "npx rimraf dist/ && npx rimraf types/ && pnpm run build"
-    },
-    "devDependencies": {
-        "@guildedjs/guilded-api-typings": "workspace:*",
-        "typescript": "4.6.4"
-    },
-    "contributors": [
-        {
-            "name": "Zaid \"Nico\"",
-            "email": "contact@nico.engineer",
-            "url": "https://github.com/zaida04"
-        },
-        {
-            "name": "Skillz4Killz",
-            "email": "guildedjs@drskillz.33mail.com",
-            "url": "https://github.com/Skillz4Killz"
-        },
-        {
-            "name": "Uhuh \"Dylan\"",
-            "email": "dylan@panku.io",
-            "url": "https://github.com/uhuh"
-        },
-        {
-            "name": "DaStormer",
-            "email": "dastormer@stormdevelopmentz.xyz",
-            "url": "https://github.com/DaStormer"
-        }
-    ],
-    "exports": {
-        ".": {
-            "require": "./dist/index.js",
-            "import": "./dist/index.mjs"
-        },
-        "./": "./"
-    },
-    "keywords": [
-        "guilded",
-        "guildedjs",
-        "guilded.js",
-        "guilded-api"
-    ],
-    "files": [
-        "dist",
-        "types"
-    ],
-    "homepage": "https://github.com/guildedjs/guilded.js/tree/main/packages/rest#readme",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/guildedjs/guilded.js.git"
-    },
-    "bugs": {
-        "url": "https://github.com/guildedjs/guilded.js/issues"
-    },
-    "gitHead": "eee38a19e0bfa812d7136cc78a6bc99e0b402b0c"
-}
