+++ packages/guilded.js/lib/colors.ts
