+++ packages/guilded.js/lib/util.ts
@@ -1,6 +1,7 @@
-import { Embed } from "@guildedjs/embeds";
 import type { RESTPostChannelMessagesBody } from "@guildedjs/guilded-api-typings";
 import { ROUTES } from "./constants";
+import { COLORS } from "./colors";
+import { Embed } from "./structures/Embed";
 
 export enum IMG_EXTENSION {
     PNG = "png",
@@ -37,3 +38,22 @@ export const buildMemberKey = (serverId: string, memberId: string): string => {
 
 export const resolveContentToData = (content: RESTPostChannelMessagesBody | string | Embed) =>
     typeof content === "string" ? { content } : content instanceof Embed ? { embeds: [content.toJSON()] } : content;
+
+/**
+ * Copyright 2015 - 2021 Amish Shah
+ * Copyrights licensed under the Apache License 2.0, https://github.com/discordjs/discord.js/blob/master/LICENSE
+ * Taken from https://github.com/discordjs/discord.js/blob/stable/src/util/Util.js#L436
+ */
+export function resolveColor(color: string | number | [number, number, number]): number {
+    if (typeof color === "string") {
+        if (color === "RANDOM") return Math.floor(Math.random() * (0xffffff + 1));
+        color = COLORS[color.toUpperCase()] ?? parseInt(color.replace("#", ""), 16);
+    } else if (Array.isArray(color)) {
+        color = (color[0] << 16) + (color[1] << 8) + color[2];
+    }
+
+    if (color < 0 || color > 0xffffff) throw new RangeError("COLOR_RANGE");
+    else if (isNaN(color)) throw new TypeError("COLOR_CONVERT");
+
+    return color;
+}
